Extract drinking water model into separate workspace package
Move the drinking water model, attributes, dataset, network wrapper,
and tests from simulation-core into a new `movici-sim-drinking-water`
package. Extract shared test fixtures into a `movici_testing` pytest
plugin so both packages can reuse them without duplication.
