+++ packages/drinking-water/pyproject.toml
@@ -1,0 +1,37 @@
+[build-system]
+requires = ["hatchling", "uv-dynamic-versioning"]
+build-backend = "hatchling.build"
+
+[project]
+name = "movici-sim-drinking-water"
+dynamic = ["version"]
+description = "Drinking water network simulation model for Movici using WNTR"
+requires-python = ">=3.10,<3.14"
+license = { text = "Movici Public License" }
+authors = [{ name = "NGinfra Movici", email = "movici@nginfra.nl" }]
+
+dependencies = [
+  "movici-simulation-core",
+  "numpy",
+  "wntr>=1.0.0",
+  "setuptools<81; python_version < '3.12'",
+]
+
+[project.entry-points."movici.plugins"]
+drinking_water = "movici_sim_drinking_water.model:Model"
+drinking_water_attributes = "movici_sim_drinking_water.attributes:DrinkingWaterNetworkAttributes"
+
+[tool.hatch.version]
+source = "uv-dynamic-versioning"
+
+[tool.uv-dynamic-versioning]
+fallback-version = "0.0.0"
+
+[tool.ruff]
+extend = "../../pyproject.toml"
+
+[tool.ruff.lint.per-file-ignores]
+"tests/*" = ["S"]
+
+[tool.ruff.lint.isort]
+known-first-party = ["movici_simulation_core", "movici_sim_drinking_water"]
