+++ packages/drinking-water/src/movici_sim_drinking_water/__init__.py
@@ -1,0 +1,5 @@
+"""Drinking water network simulation model using WNTR"""
+
+from .model import Model
+
+__all__ = ["Model"]
