+++ packages/drinking-water/src/movici_sim_drinking_water/attributes.py
