+++ packages/drinking-water/src/movici_sim_drinking_water/json_schemas/drinking_water.json
