+++ packages/simulation-core/pyproject.toml
@@ -54,9 +54,6 @@ models = [
   "pyproj>=3.0.1",
   "movici-geo-query>=1.1.1",
   "scipy",
-  "wntr>=1.0.0",
-  # setuptools<81 supplies pkg_resources, required for wntr on older python versions
-  "setuptools<81; python_version < '3.12'",
 ]
 sqlite = ["sqlalchemy>=1.4,<3.0"]
 all = [
@@ -65,7 +62,6 @@ all = [
   "pyproj>=3.0.1",
   "movici-geo-query>=1.1.1",
   "scipy",
-  "wntr>=1.0.0",
   "sqlalchemy>=1.4,<3.0",
 ]
 
@@ -103,6 +99,9 @@ traffic_kpi = "movici_simulation_core.models.traffic_kpi.model:Model"
 udf = "movici_simulation_core.models.udf_model.udf_model:UDFModel"
 unit_conversions = "movici_simulation_core.models.unit_conversions.model:Model"
 rules = "movici_simulation_core.models.rules.model:Model"
+
+[project.entry-points.pytest11]
+movici_testing = "movici_simulation_core.testing.fixtures"
 
 [tool.hatch.version]
 source = "uv-dynamic-versioning"
