--- packages/simulation-core/src/movici_simulation_core/models/drinking_water/__init__.py
@@ -1,5 +1,0 @@
-"""Water network simulation model using WNTR"""
-
-from .model import Model
-
-__all__ = ["Model"]
